realtek: rtl930x: Add support for Zyxel XGS1010-12
The Zyxel XGS1010-12 is an 8x 1 GBit, 2x 2.5 GBit NBaseT ports and
2 SFP+ ports switch. It is identical to the Zyxel XGS1210-12, except for
the partition layout which is probably caused by the basic firmware
coming from the SDK with the default dual partition layout. To ensure
backwards compatibility and more importantly to those who also do not
wish to replace their bootloader, a second image type with the postfix
`_vendor` is also added. Using our custom U-Boot adds about ~7MiB of
storage to the firmware image compared to stock.

In addition to the above, the XGS1010-12 lacks a physical reset button,
but all circuitry is in place to support this, including the whole in
the metal housing. If a switch is added at a later time, the switch is
basically converted to an XGS1210-12 and thus that firmware should be
used instead.

As this switch does not feature any form of management or firmware
updateability, one must use the serial UART either through XYZ-modem
(untested) or tftp. The instructions to do so are identical as with the
XGS1250-12.

When using our customized U-Boot, `upgrade runtime` can be used
instead over ftp as well.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>