realtek: clk: Reorganize rate recalculation
To better adapt to future devices reorganize the register reading. A small
macro helps us to build a switch statement for SOC and clock, making it
easier to build a matrix. This does cause a lot of code duplication
however.

In the future, it probably makes more sense to split these functions up
to a 'per soc' style.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>
Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>