realtek: clk: Access ctrl with full address
The RTl930x has it's address space more cluttered and keeps the mem
controller in the SoC addressing section. To allow for easier addressing
and little code-changes, access or clock controller registers with the
full address.

In the future we ought to get all these addresses from the devicetree,
but for now, this will have to do.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>