realtek: Do not make distinction on SRAM objects
The SRAM object source code has include guards to prevent them from
causing problem wrong architectures and as a side-effect reduce code
size (as both object always exist in the source).

When going towards a 'unified kernel' approach this is not possible
(only runtime distinctions) and thus, we have to remove these compile
time guards.

This does increase the kernel size by a probably 1 kib. If we want to
avoid this micro-size optimization, we can move these objects into the
filesystem and load them on-demand via the firmware loading methodology
as in the end, these are really firmwares in a sense.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>