realtek: Move to unified target configuration
There doesn't seem to be much of a technical reason, to have multiple
sub-targets for our 'otto' target. So lets combine them into one unified
configuration. The default config is the merger of the previous ones.

While this allows for a shared kernel binary at the least, this is
of course not yet the case, but does open up the possibility of that and
even having a single firmware file for all our realtek targets in the
future.

Note, that RTL931X is mostly excluded and still keeps its own
configuration option, as there is some boot-time weirdness with the
load-address going on that needs some extra attention.

This might somewhat revert commit fce11f6849 ("realtek: Create 4 different
Realtek Platforms").

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>