realtek: Do not parse `ls` output
It is considered very bad practise to parse the output of `ls` in
shell scripts. We can do it easily without, however we do get an
unsorted list. Due to the lack of 'human sort' in the default OpenWRT
busybox sort function, we have to hack a little bit around it.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>