realtek: Be more versatile in getting the MAC address
Most of the realtek targets get the MAC address/range pretty standard
places. The U-Boot environment or the U-Boot 'sys' environment. Instead
of trying to read directly from flash, get these things from variables
if possible. Only if this fails, we fall back to alternative means.

This also means that this is generic enough, to support most/all of our
switches, except the HP ones, which embed the MAC address weirdly in the
header.

This means that, we first try to get the MAC address from the
devicetree, potentially set by U-Boot.

Secondly, we check if we have the U-Boot env tools to ask both the
environments, which gives us a nice and clean way to get this data.

If this fails, we'll resolve to more aggressive methods by dumping the
data from the various flashes. We are more flexible in naming of the
second partition, which could be named either `env2` or `sysenv`. This
more aggressive method is also useful for initramfs mode, which usually
does not expose the environment. A little bit of a hack ;).

Finally, if we fail to get a valid MAC address, resort to a random MAC
address supplied by OpenWRT.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>