base-files: Use sed instead of grep and awk
We can use a regex (with sed) to obtain the mtd index instead of grep
and awk, which a sub-shell spawned less and a pipe less.

However, performance is not really the reason for this change. This
change, especially the global regex definition, allows us to also obtain
the partition size, erase block size to allow us to generate the
environment partition more generically.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>