realtek: Add MAC control defines
As we have gained much more knowledge, we know in more detail a lot of
the register fields. This commit adds defines for the currently known
ones. "Guesses" are left as comments until we can _prove_ that these
fields are correct.

Some fields are duplicated, which will be removed in a following patch.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>