realtek: Fix duplex check on rtl83xx
The state variable 'duplex' only contains the value 'DUPLEX_FULL' or
'DUPLEX_HALF' which will resolve to either 1 or 0 respectively.

Comparing however to RTL838x_DUPLEX_MODE, the register flag which sits
at bit 3, will never compare true, as we'd be comparing '0 == 3' or '1
== 3'.

It is a bit puzzeling how this is working currently however.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>