realtek: Split MAC related switch ops
The mac port control, force mac controller and mac handling registers are
all just slightly different between the various platforms. Cleaning these up,
would result in a function with huge switch case (or if else) in them as
each platform has their own branch.

While ideally, we'd have one function that easily handles all platforms,
our understanding currently is not good enough to be able to accomplish
this today. In the future, we might combine these functions into more
elegant ways.

We probably will replace this in the future with regmaps and the likes,
once we have better support and understanding in place.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>