realtek: eth: Really enforce RX length
The driver currently wants to set the maximum receive length to 1600/512
bytes. Testing shows that for RTL83XX this does not work because the
enable bit is missing.

Add the missing enable bits.

Note, that investigation is still needed if the TX padding en bit
should be set at all here.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>