realtek: eth: Do not manually update counters anymore
The current code mentions about a bug of RTL838X SOCs in the receive path.
Intensive testing shows that counter increasing/decreasing might miss
some updates and space in buffer is leaking slowly. Nevertheless the
current workaround is a modification of the counter register.

With the newly introduced rtl83xx counter update routines we can now
remove the double coding. The calling location is changed to the beginning
of the receive path. This is no problem because:

- On RTL83XX the counters are much smaller than the ring size
- On RTL93XX the counters are ring size minus 2

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>