realtek: rtl930x: add support for TP-LINK TL-ST1008F
The TP-LINK TL-ST1008F is an 8-switch with 10 SFP+ cages capable of
between 10Mbps and 10Gbps depending on the installed transceiver.
It is sold as an unmanaged switch, but has 32MiB flash and 512MiB of RAM
which is ample for running mainline Linux on it.

Networking is broken on stock U-Boot and this being an unmanaged switch
it does not have a management interface to install updated firmware, so
one has to download the firmware initially over serial using ymodem or
similar. The main reason networking does not work is that TP-LINK uses
an I/O expander to drive the TX_DIS pin on the SFP+ transceiver which
U-Boot has not been extended to support. It is unknown if this is the
only issue preventing U-Boot networking from working.

Signed-off-by: Lorenz Brun <lorenz@brun.one>