realtek: Make mach header more generic
The mach header has been abused for all our platforms over the years. It
is no longer specific to the rtl83xx series as the name suggests, but
for the whole otto (as realtek names it internally) platform.

Lets rename it as such, while doing this, also cleanup all header files
to properly include this header.

The big todo here is, to remove all of this in the future, as nobody
should be touching or using this header anyways.

A small side-effect, with all this cleanup, we no longer need to set the
search path for this header globally and thus, instead of renaming, can
just remove the entry all together.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>