realtek: Add ZBOOT debugging support
When in a tight spot, getting something out on the serial UART is
crucial. It is the last thing before going down the JTAG rabbit hole.

This is never intended to be enabled in production or even regular
(development builds) and is only to be used for doing _very_ early
kernel debugging.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>