realtek: Store number of LEDs in devicetree
Each port needs two LED-related settings. Which LED-set to use and how
many of LEDs are connected to a port. Since these are related, but not
coupled, we can't use any algorithm to figure this out otherwise, as a
LED-set definition could account for 4 LEDs, but a port may only wire 2
of them. E.g. a switch with 8 ports 4 with POE 4 without, which could
mean 4 ports with 4 LEDs (2 LAN LEDs, 2 POE LEDs) and 4 ports with just
2 LEDs (lan).

To be more future-proof, we store this as a bitmap, so that in theory,
we could determine 'led 3 and led 0' are connected. The rtl930x hardware
doesn't seem to support this however. It is trivial to convert this to
an actual number of leds with the `hweightN` series of bitop functions.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>