WIP: Add whitelabel 24p-4s+r1 switch
The Whitelabel ws-sw-24p-4s+r1 switch is a 24x 1 Gbit and 4x SFP+ ports
switch with POE.
Two images are offered, one following the vendor layout of the
partitions, and one 'optimized' for OpenWRT, where all flash space is
used. A backup of the full flash is strongly recommended.

Hardware:
        - rtl9301 SoC: MIPS 34Kc @ 800MHz
        - 128 MiB DRAM
        - 32MBit NOR Flash W25Q256JV
        - 3x RTL8218D 8x GBit PHY
        - 4x SFP+ Slots
        - 1 Reset button
        - 1 System LED
        - Externally accessable cisco-style console UART

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>