realtek: dts: Move user I/O to rtl930x main dtsi
The LED and reset button is 'the same' for all devices that support
(populate/route) them. There is only a single LED controller in the
RTL930x SoC and if used, it is used for exactly this one purpose.

The same is the reset switch, if it is used/populated, it behaves the
same.

Lets add them to the main/common dts file for easier re-use.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>